+++ float_masking.py
@@ -1,14 +1,18 @@
 import pandas as pd
-import random
-import string
 import struct
 import numpy as np
 # Load the input Excel file
-input_file = r'C:\Users\sethir919\Desktop\project\masking-input.xlsx'
+input_file = r'C:\Users\sethir919\Desktop\masking project\Data-Obfuscation\data\masking-input.xlsx'
 data = pd.read_excel(input_file)
 
+def chain(start, *funcs):
+    res = start
+    for func in funcs:
+        res = func(res)
+    return res
+
 # Define a function to mask a single name
-def mask_name(weight):
+def mask_float_by_str(weight): #56.64
     samp=str(weight)
     one = samp[0]
     two=samp[1]
@@ -16,22 +20,10 @@ def mask_name(weight):
     four= samp[-1]
     masked_name = three+four+"."+one+two
     return  masked_name 
+#data['Weight']=data['Weight'].apply(mask_float_by_str)
+#excel=print(data['Weight'][0])
 
-
-def mask_float_values(input_list, mask_char='*', mask_start=0, mask_end=None):
-    """
-    Mask float values in input_list with mask_char from mask_start to mask_end indices.
-    If mask_end is not provided, it defaults to the end of the input_list.
-    """
-    if mask_end is None:
-        mask_end = len(input_list)
-    masked_list = input_list.copy()
-    for i in range(mask_start, mask_end):
-        if isinstance(masked_list[i], float):
-            masked_list[i] = mask_char
-    return masked_list
-
-def mask_float(value):#3,.5
+def mask_float_by_precision(weight):
     """
     Mask a float value by replacing the last `precision` digits with `mask_value`.
     :param value: the float value to be masked
@@ -39,18 +31,21 @@ def mask_float(value):#3,.5
     :param precision: the number of digits to be masked
     :return: the masked float value
     """
-    precision=8
-    factor = 10 * precision #20 
-    mask_value=7*int(value * factor)
-    masked_value = int(value * factor) # get the integer part before the decimal
+    precision=3
+    factor = 2 * precision #20 
+    mask_value=7*int(weight * factor)
+    masked_value = int(weight * factor) # get the integer part before the decimal
     rand=str(mask_value) * precision
     masked_value = round(float(str(masked_value) + '.' + rand),2) # append the mask_value
-    return masked_value%100
+    res=masked_value%100
+    return res
+#data['Weight']=data['Weight'].apply(mask_float_by_precision)
+#print(data['Weight'])
 
 # define two floating-point numbers
-def mask_val(val):
-    a =76.89
-    b =val
+def mask_float_by_add(weight):
+    a =weight
+    b =3.14
     # convert the numbers to their IEEE 754 binary representation
     a_bin = struct.pack('!f', a)
     b_bin = struct.pack('!f', b)
@@ -60,22 +55,27 @@ def mask_val(val):
 # convert the result back to a floating-point number
     result = struct.unpack('!f', result_bin)[0]
     return result
+#data['Weight']=data['Weight'].apply(mask_float_by_add)
 
 
-def mask_or(val):
+def mask_float_by_or(weight):
     a =150
-    b = 200
+    b =weight
     # convert the numbers to their IEEE 754 binary representation
     a_bin = struct.pack('!f', a)
     b_bin = struct.pack('!f', b)
-    # perform a bitwise AND operation on the binary strings
+    # perform a bitwise OR operation on the binary strings
 
     result_bin = bytearray([x | y for x, y in zip(a_bin, b_bin)])
 # convert the result back to a floating-point number
     result = struct.unpack('!f', result_bin)[0]
     return result
+#data['Weight']=data['Weight'].apply(mask_float_by_or)
 
-def float_bit_shift_and_mask(x, shift_amount, mask):
+def mask_float_by_shift(x):
+    # Define the shift amount and mask
+    shift_amount = 1
+    mask = 1111000011000
     # Convert the floating-point number to its IEEE 754 binary representation
     x_bin = bytearray(struct.pack('!f', x))
 
@@ -90,15 +90,9 @@ def float_bit_shift_and_mask(x, shift_amount, mask):
 
     # Return the resulting floating-point number
     return result
+#data['Weight']=data['Weight'].apply(mask_float_by_shift)
 
-# Define the shift amount and mask
-shift_amount = 1
-mask = 1111000011000
-
-#data['Weight'] = data['Weight'].apply(float_bit_shift_and_mask, shift_amount=shift_amount, mask=mask)
-#print(data['Weight'])
-
-def float_rotate(x):
+def mask_float_by_rotate(x):
     r=150
     # Convert float to binary string
     bits = bin(struct.unpack('!I', struct.pack('!f', x))[0])
@@ -109,10 +103,11 @@ def float_rotate(x):
     # Convert binary string back to float
     x_rotated = struct.unpack('!f', struct.pack('!I', int(bits, 2)))[0]
     return round(x_rotated,2)
-#data['Weight']=data['Weight'].apply(float_rotate)
+#data['Weight']=data['Weight'].apply(mask_float_by_rotate)
 #print(data['Weight'])
 
-def swap_floats_one(a,b):
+def mask_float_by_swap1(a):
+    b=77
     """
     Swaps the integer parts of two float values and returns them as a tuple.
     """
@@ -120,42 +115,16 @@ def swap_floats_one(a,b):
     b_int, b_dec = divmod(b, 2)
     return b_int + a_dec
 # Apply the swap_floats() function to the 'Weight' column
-# data['Weight'] = data['Weight'].apply(swap_floats_one,b=77)
+#data['Weight'] = data['Weight'].apply(mask_float_by_swap1)
 
 
-def swap_floats(a):
+def mask_float_by_swap2(weight):
     """
     Swaps the integer parts of two float values and returns them as a tuple.
     """
-    a_int, a_dec = divmod(a, 2)
+    a_int, a_dec = divmod(weight, 2)
     return a_dec +a_int
-# Apply the swap_floats() function to the 'Weight' column
-#data['Weight'] = data['Weight'].apply(swap_floats)
-
-
-'''def shuffle_column(weight):
-    shuffled = np.random.permutation(weight)
-    return shuffled
-data['Weight'] = data['Weight'].apply(shuffle_column)
-print(data['Weight'])'''
-import pandas as pd
-import numpy as np
-
-def shuffle_column_values(df, column_name):
-    # shuffle the values in the specified column
-    shuffled_column = df[column_name].sample(frac=1).reset_index(drop=True)
-    # create a new DataFrame with shuffled column
-    shuffled_df = pd.DataFrame({
-        column_name: shuffled_column})
-    return shuffled_df
-shuffled_weight = shuffle_column_values(data, 'Weight')
-print(shuffled_weight)
-
-
-
-
-
-
+#data['Weight']=data['Weight'].apply(mask_float_by_swap2)
